music: move EmbeddedArt and EmbeddedArtInfo from MusicInfoTag.h into its own header file
